Remove normalize fn for type, return default props instead
