+++ packages/dataviews/src/components/dataviews-picker/index.tsx
@@ -29,7 +29,7 @@ import DataViewsViewConfig, {
 	DataviewsViewConfigDropdown,
 	ViewTypeMenu,
 } from '../dataviews-view-config';
-import normalizeFields from '../../field-types/normalize-fields';
+import normalizeFields from '../../field-types';
 import type { ActionButton, Field, View, SupportedLayouts } from '../../types';
 import type { SelectionOrUpdater } from '../../types/private';
 type ItemWithId = { id: string };
