+++ packages/dataviews/src/field-types/date.tsx
@@ -74,40 +74,38 @@ const sort = ( a: any, b: any, direction: SortDirection ) => {
 	return direction === 'asc' ? timeA - timeB : timeB - timeA;
 };
 
-export default function normalizeField< Item >(): TypeProvidedProps< Item > {
-	return {
-		type: 'date',
-		render,
-		Edit: 'date',
-		sort,
-		isValid: {
-			elements: true,
-			custom: () => null,
-		},
-		enableSorting: true,
-		enableGlobalSearch: false,
-		defaultOperators: [
-			OPERATOR_ON,
-			OPERATOR_NOT_ON,
-			OPERATOR_BEFORE,
-			OPERATOR_AFTER,
-			OPERATOR_BEFORE_INC,
-			OPERATOR_AFTER_INC,
-			OPERATOR_IN_THE_PAST,
-			OPERATOR_OVER,
-			OPERATOR_BETWEEN,
-		],
-		validOperators: [
-			OPERATOR_ON,
-			OPERATOR_NOT_ON,
-			OPERATOR_BEFORE,
-			OPERATOR_AFTER,
-			OPERATOR_BEFORE_INC,
-			OPERATOR_AFTER_INC,
-			OPERATOR_IN_THE_PAST,
-			OPERATOR_OVER,
-			OPERATOR_BETWEEN,
-		],
-		getFormat,
-	};
-}
+export default {
+	type: 'date',
+	render,
+	Edit: 'date',
+	sort,
+	isValid: {
+		elements: true,
+		custom: () => null,
+	},
+	enableSorting: true,
+	enableGlobalSearch: false,
+	defaultOperators: [
+		OPERATOR_ON,
+		OPERATOR_NOT_ON,
+		OPERATOR_BEFORE,
+		OPERATOR_AFTER,
+		OPERATOR_BEFORE_INC,
+		OPERATOR_AFTER_INC,
+		OPERATOR_IN_THE_PAST,
+		OPERATOR_OVER,
+		OPERATOR_BETWEEN,
+	],
+	validOperators: [
+		OPERATOR_ON,
+		OPERATOR_NOT_ON,
+		OPERATOR_BEFORE,
+		OPERATOR_AFTER,
+		OPERATOR_BEFORE_INC,
+		OPERATOR_AFTER_INC,
+		OPERATOR_IN_THE_PAST,
+		OPERATOR_OVER,
+		OPERATOR_BETWEEN,
+	],
+	getFormat,
+} satisfies TypeProvidedProps< any >;
