+++ packages/dataviews/src/field-types/media.tsx
@@ -3,20 +3,18 @@
  */
 import type { TypeProvidedProps } from '../types/private';
 
-export default function normalizeField< Item >(): TypeProvidedProps< Item > {
-	return {
-		type: 'media',
-		render: () => null,
-		Edit: null,
-		sort: () => 0,
-		isValid: {
-			elements: true,
-			custom: () => null,
-		},
-		enableSorting: false,
-		enableGlobalSearch: false,
-		defaultOperators: [],
-		validOperators: [],
-		getFormat: () => ( {} ),
-	};
-}
+export default {
+	type: 'media',
+	render: () => null,
+	Edit: null,
+	sort: () => 0,
+	isValid: {
+		elements: true,
+		custom: () => null,
+	},
+	enableSorting: false,
+	enableGlobalSearch: false,
+	defaultOperators: [],
+	validOperators: [],
+	getFormat: () => ( {} ),
+} satisfies TypeProvidedProps< any >;
