+++ packages/dataviews/src/field-types/telephone.tsx
@@ -28,31 +28,29 @@ const sort = ( a: any, b: any, direction: SortDirection ) => {
 	return direction === 'asc' ? a.localeCompare( b ) : b.localeCompare( a );
 };
 
-export default function normalizeField< Item >(): TypeProvidedProps< Item > {
-	return {
-		type: 'telephone',
-		render,
-		Edit: 'telephone',
-		sort,
-		isValid: {
-			elements: true,
-			custom: () => null,
-		},
-		enableSorting: true,
-		enableGlobalSearch: false,
-		defaultOperators: [ OPERATOR_IS_ANY, OPERATOR_IS_NONE ],
-		validOperators: [
-			OPERATOR_IS,
-			OPERATOR_IS_NOT,
-			OPERATOR_CONTAINS,
-			OPERATOR_NOT_CONTAINS,
-			OPERATOR_STARTS_WITH,
-			// Multiple selection
-			OPERATOR_IS_ANY,
-			OPERATOR_IS_NONE,
-			OPERATOR_IS_ALL,
-			OPERATOR_IS_NOT_ALL,
-		],
-		getFormat: () => ( {} ),
-	};
-}
+export default {
+	type: 'telephone',
+	render,
+	Edit: 'telephone',
+	sort,
+	isValid: {
+		elements: true,
+		custom: () => null,
+	},
+	enableSorting: true,
+	enableGlobalSearch: false,
+	defaultOperators: [ OPERATOR_IS_ANY, OPERATOR_IS_NONE ],
+	validOperators: [
+		OPERATOR_IS,
+		OPERATOR_IS_NOT,
+		OPERATOR_CONTAINS,
+		OPERATOR_NOT_CONTAINS,
+		OPERATOR_STARTS_WITH,
+		// Multiple selection
+		OPERATOR_IS_ANY,
+		OPERATOR_IS_NONE,
+		OPERATOR_IS_ALL,
+		OPERATOR_IS_NOT_ALL,
+	],
+	getFormat: () => ( {} ),
+} satisfies TypeProvidedProps< any >;
