+++ packages/dataviews/src/utils/filter-sort-and-paginate.ts
@@ -36,7 +36,7 @@ import {
 	OPERATOR_IN_THE_PAST,
 	OPERATOR_OVER,
 } from '../constants';
-import normalizeFields from '../field-types/normalize-fields';
+import normalizeFields from '../field-types';
 import type { Field, View } from '../types';
 
 function normalizeSearchInput( input = '' ) {
